Move toPbjTimestamp() and fromPbjTimestamp() to PbjConverters in consensus-module.
