+++ platform-sdk/consensus-model/src/main/java/org/hiero/consensus/model/hashgraph/ConsensusRound.java
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.consensus.model.hashgraph;
 
+import static org.hiero.consensus.model.PbjConverters.fromPbjTimestamp;
+
 import com.hedera.hapi.node.state.roster.Roster;
 import com.hedera.hapi.platform.state.ConsensusSnapshot;
 import com.swirlds.base.utility.ToStringBuilder;
@@ -13,7 +15,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 import org.hiero.base.iterator.TypedIterator;
-import org.hiero.base.utility.CommonUtils;
 import org.hiero.consensus.model.event.CesEvent;
 import org.hiero.consensus.model.event.ConsensusEvent;
 import org.hiero.consensus.model.event.PlatformEvent;
@@ -190,7 +191,7 @@ public class ConsensusRound implements Round {
      */
     @Override
     public @NonNull Instant getConsensusTimestamp() {
-        return Objects.requireNonNull(CommonUtils.fromPbjTimestamp(snapshot.consensusTimestamp()));
+        return Objects.requireNonNull(fromPbjTimestamp(snapshot.consensusTimestamp()));
     }
 
     /**
