+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/TransactionFactory.java
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.hiero.otter.fixtures;
 
+import static org.hiero.consensus.model.PbjConverters.toPbjTimestamp;
+
 import com.google.protobuf.ByteString;
 import com.hedera.hapi.platform.event.StateSignatureTransaction;
 import com.hedera.node.app.hapi.utils.CommonPbjConverters;
@@ -8,7 +10,6 @@ import com.hederahashgraph.api.proto.java.Timestamp;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.time.Instant;
 import java.util.List;
-import org.hiero.base.utility.CommonUtils;
 import org.hiero.consensus.model.node.NodeId;
 import org.hiero.otter.fixtures.network.transactions.EmptyTransaction;
 import org.hiero.otter.fixtures.network.transactions.HashPartition;
@@ -46,7 +47,7 @@ public class TransactionFactory {
      */
     @NonNull
     public static OtterTransaction createFreezeTransaction(final long nonce, @NonNull final Instant freezeTime) {
-        final Timestamp timestamp = CommonPbjConverters.fromPbj(CommonUtils.toPbjTimestamp(freezeTime));
+        final Timestamp timestamp = CommonPbjConverters.fromPbj(toPbjTimestamp(freezeTime));
         final OtterFreezeTransaction freezeTransaction =
                 OtterFreezeTransaction.newBuilder().setFreezeTime(timestamp).build();
         return OtterTransaction.newBuilder()
