+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/state/service/SnapshotPlatformStateAccessor.java
@@ -2,7 +2,7 @@
 package com.swirlds.platform.state.service;
 
 import static java.util.Objects.requireNonNull;
-import static org.hiero.base.utility.CommonUtils.fromPbjTimestamp;
+import static org.hiero.consensus.model.PbjConverters.fromPbjTimestamp;
 
 import com.hedera.hapi.node.base.SemanticVersion;
 import com.hedera.hapi.platform.state.ConsensusSnapshot;
